progress: 35.81% of code matched (5058/8216 functions, 5293/6376 objects)
Source-commit: 0de3f8da7232b7453764afcae36559eb7e0d6950

+++ src/game/game_fn_800A3104.c
@@ -8,15 +8,9 @@ typedef struct Object800A3104 {
 
 extern int fn_800A30C0(Object800A3104*);
 
-/*
- * Behavior-complete reconstruction. The retail TU inlined fn_800A30AC here
- * and retained its result in r0 before moving it to r31. With this function
- * split into an independent TU, MWCC coalesces that value directly into r31,
- * leaving the generated body four bytes shorter.
- */
 int fn_800A3104(Object800A3104* object, int enabled)
 {
-    u16 old_value = object->flags & 1;
+    int old_value = (u16)(object->flags & 1);
 
     if (fn_800A30C0(object) == 0) {
         if (enabled != 0) {
